Upgrading the TypeScript files
This splits up the index file into individual files and fixes some type
issues with React type definition upgrades.
