+++ src/component.ts
@@ -1,0 +1,9 @@
+import Mixin = require("././mixin");
+
+class Component<P, S> extends Mixin<P, S> implements React.CompositeComponent<P, S> {
+    render(): React.ReactElement<any> {
+        return null;
+    }
+}
+
+export = Component;
